Refactored DS18B20 sensor code out of main, tmp uart msgs added
